Fix failing test
Test fails due to UK constraint error. This is because we have
an explicit insert into a table and then when trying to insert
using the sequence, we have a collision. This was not a problem
in local testing as the sequence had advanced past the value '1',
but on a fresh DB instance we encounter this problem.