Add ansible http_server role
Adds a role based on the lobby role to install and run http_server.
This update also cleans up the lobby role of unused properties
that are exclusive to http_server now.

Note: not well tested, there may be some issues in this config, for now
this lays down the obvious boiler plater in preparation for a more serious
effort to ensure that http server is deployed and working on at least
prerelease.