Remove unused BadWordController#add method (#4854)
Bad words are added directly to DB (for better or worse), the method to add a bad word
is currently only used in test code.

Note: we introduce a test gap to verify that usernames do not contain 'bad words'. For now this is intentional as DB tests are in progress of being updated to use database-rider. With further work in the DB area with luck we'll fix this testing gap soon enough.