Parameterizer retry count and Use a feign custom JSON encoder to better support String encoding
_Parameterized retry count_

This is added mainly so that test can reduce retries and speed up tests. When verifying error
cases we do not need to retry when we expect all attempts to fail.

_Custom JSON Encoder_

Implementing a work-in-progress future http client that sends simple String parameters,
the existing encoder will add surrounding quotes around these strings values. This means
that instead of a server receiving for example "value", it will get "\"value\"".

The custom JSON encoder is a "pass-thru" to skip encoding when we are encoding a string
encoder, otherwise we use the (existing) GSON encoder to encode objects. This means string
values are simply encoded as-is, and objects are otherwise rendered into JSON.