Add JTableBuilder
Utility class to provide a builder-like API for constructing a JTable.
Purpose of the builder is to hide Swing API details and simplify the API
that the rest of the code would use.

Note: table builder is not yet used, but is planned to be used extensively for
moderator toolbox-2.0 to display database table data.