Add ModeratorToolbox http clients for API key validation and bad word management
This update adds http clients that will be used for the moderator-toolbox-2.0.
The initial features supported are for API key validation and for bad word management
(CRUD operations for the bad-word table).

Currently this http client is not currently used outside of in-flight work that is not
yet checked in.

One important design decision taken is that all moderator toolbox operations are grouped
under one http client. The controller methods servicing these methods will on the other
hand be broken up by moderator toolbox tab. For example there will be a dedicated controller
for API key validation and another for bad-word management, and as a further example
yet another server controller for user ban management.

API-key note:
- The http client and server is designed so that we will pass an api-key from client to the server.
It is intended that we will issue each moderator a unique and lengthy api-key that will they
enter into the TripleA UI on an admin screen that will then be saved in client settings. This key
will then be sent to the server for any moderator related actions. The server side will for any
moderator endpoint verify this key and do rate limiting to ensure brute-force attacks cannot occur.