Bring in Toast.java class to show pop-up toasts
Toast.java is a copy/paste from: https://www.geeksforgeeks.org/java-swing-creating-toast-message/
It has been modified slightly to fit our coding style and for some minor improvement like parameterized
window height and width and some spelled out variable names.

To make use of this class, MessagePopup.java is introduced which provides a simplified API to launch
the toast window.

The toast window will eventually be used with moderator-toolbox-2.0 to show action confirmations. For example
when values are added or removed from database the confirmation message will be in a toast message rather
than a confirmation dialog.

For reference, Toast is a UI widget, a small window that appears that is not clickable and pop-ups and then
slowly fades away informing a user of a message.