Add copy/pasted ButtonColumn.java
This class can be used to convert a column in a JTable to be buttons.
The button listener can be customized and is provided the row number
and a table model which can be then be used to manipulate the table.

Of note, the button text is defined by the value that was previously
in the the table cell.

This class is largely a copy/paste from:  http://www.camick.com/java/source/ButtonColumn.java
There are some very minor updates to make it conform to our checkstyle rules and is
otherwise largely unmodified.