Lobby-db-dao consolidation and new sub-project
This update adds a COPY of DAO classes to a new sub-project lobby-db-dao.
Most of this code is *not* new.

Notable additions/changes:
- ModeratorAuditHistoryController is updated to use JDBI.
- BadWordsDao is updated to include methods that will be used for moderator-toolbox

Existing code is not yet altered to make use of the new classes, so the new
sub-project is currently unused except for in-flight work that is not yet checked in.

_Motivation_
There is overlap between http-server needing the moderator audit history and the java lobby.
To avoid duplication of enums and other common functionality, this PR begins the process
of creating a common sub-project that is responsible for all DAO.

_JDBI migration note_
JDBI really simplifies JDBC code, with database-rider we do not need to update any tests.
It is intended for eventually all the DAO classes named *Controller to be moved here,
tested with database-rider and converted to JDBI.