Fix datasets to remove data from tables with FKs
One downside of database-rider datasets is that tables with FKs are not automatically cleared.
Any new table with a FK needs to be added to any existing dataset where the primary key exists.
To explain that, if we have table A, and then a new table B with FK to A, then any existing dataset
that uses table A will delete it. If we have run a test that populates table B, then the deletion of
table A will cause a FK constraint violation. This introduces a non-determinism where tests can
fail simply depending on the order in which they are run.

This update fixes this whooopsie and deletes tables that have FKs into lobby_user
(notably the moderator api key table).

Second, integration.yml is simply not used and is being dropped.