Support for moderator single-use-key
Adds support and concept for a single-use-key that moderators can use to register a
password and be granted a permanent API key that is stored on their system.

This allows for keys to be distributed and then a new one to be regenerated that
is not known to anyone other than the moderator.

Change Summary:
- incorporate logic so that a key password is registered and hashed with the main moderator key
- update UI to prompt for single-use-key and password if no API key is stored in client settings
- Update toolbox flow to always at least ask for api key password
- add DB table for single-use-key
- update backend to do validation of api key salted with password