Remvoe checkStyle rules that conflict with Google Java Format
Remove rules from checkstyle that conflict with google java format and
are enforced by formatter (spotlessCheck fails if these are violated).

When moving to google java format, some of these rules fail and are simply
not appropriate for the new formatting spec. Again these rules are not needed
as the formatter will enforce the formatting and spotlessCheck will not pass
if the formatting is wrong.