Trim requested user names
Leading or trailing whitespace on requested user names is chopped off with this update. Avoids a user from
creating a same nickname by adding spaces to the end of an existing name.

Typical clients cannot add whitespace to the end of a username, disallowed by UI. If a motivated user were
to create a custom/hacked client, then they could send such a username.