Move add/remove error listener from IMessenger to IClientMessenger
Goal is to remove calls that result in no-op. ServerMessenger and LocalNoOpMessenger are no-ops for add/remove error listener. The methods are moved to related interfacer IClientMessenger where is always an effect. Cases where those methods are called on IMessenger, and sometimes when we have an IClientMessenger, an 'instanceof' check is added to cast the IMessenger to an IClientMessenger. In cases when the IMessenger is always a ServerMessenger (and a no-op), those calls are removed.