Update pull request template
Make the pull request a bit more self directed with the goal that relevant sections would be filled in.
Add category list to make it a bit easier to define what kinds of changes are in the PR.
Add bug fix questions to encourage that the root cause be explained and help ensure we are thinking deeply about bugs and their fixes.