Remove checkstyle rules that conflict with google format
Google format will conditionally place right curly on same line when there is empty content for: while loops, for loops, class def, method def

This update removes the checkstyle rule that checks for right curly to be on its own line which conflicts with the above cases. The google-format will correct this and spotless gradle plugin will enforce this.