Update README links now that docs are on wiki
(1) license file is moving to LICENSE.md
(2) Uptime robot link can live on wiki only, does not need to be on readme
(3) Remove gitter link, our response to gitter is slow and we do not need to advertise that to the public.
(4) Update documentation link to be to the wiki.