Unit scroller improvements (#5039)
* Unit scroller improvements

(1) Notify when all units moved via popup. If this is annoying to users, it can be turned off with a 'do not show again' button or turned off/on via game settings.
(2) If all units in current territory are moved, 'c' will focus on next territory instead of focusing on the current (and empty) territory
(3) Sleep only movable units in a territory. For example landing bombers with an AA gun, it's common to only want the AA gun to be slept.