Fix double casualties being shown
Bisected the error to refactoring update: e2be5aa16f6da7602a1802f62aa2bcfa3188bb71
The return value of `startGame`, an `IDisplay` is added to `IGame`. In the refactor, this add is dulicated and extracted leading to two calls to `IGame.addDisplay(display)`. This then doubles messages sent to the display leading to double casualties to be displayed.

The fix is to remove the extra `addDisplay` call.