Disallow login with name that is already logged in
Add a case insensitive check that a player name does not match any that are logged in. For example, if player "TEST" is logged, the name "test" would be rejected. The check for registered user name is case insensitive, it should then not be possible to use this to block registered users from logging in.