Simplify Version and set to 2.0.0
Set version to 2.0.0; Fits with new version criteria where first value is engine compatibility, second is compatible release number, third is build number. See: https://github.com/triplea-game/triplea/issues/4875 for more details and discussion.

(1) The 'lesserThan' API usage can be replaced by using 'GreaterThan' and swapping argument positions

(2) Remove '@buildId@' replacement in Gradle and use '0' for the dev build. This can be done by having travis overwrite the properties file with the build number, at which point gradle can read the property value directly without doing a token replacement. Meanwhile we can set the local version number to be '2.0.0' and then local builds do not need to do a token replacement to 'dev'. Since the last number is the build number is always increasing, a '0' for the build number clearly indicates a 'dev' build and we do not need to use a special 'dev' string to represent it.

(3) Simplify parsing of Version, use 'splitter' instead of a regex

(4) Use lombok to simplify Version getters and Setters, equals and hashcode

(5) Remove EngineVersion wrapper and move those methods to Version itself