Use invokeLater instead of invokeAndWait
Bit of a blind fix to try and avoid deadlock in BattlePanel with 'invokeAndWait'.

We are seeing deadlock happen with a call to 'setVisible' from within EDT while another thread is calling invokeAndWait. With luck by calling 'invokeLater' instead, the 'setVisible' call will be queued to be called after the other thread executes invokeAndWait.