Add email to password forgot and move to http-server
(1) For forgot password, request user enter their email in addition to their username
- This is so we can more truly avoid fake password requests. A users email is not known publicly while username is. By requesting both username and email, we are requesting a non-public piece of information that only the user should know.
- We need to also ask for username in additional to email as emails are not unique in TripleA but usernames are.

(2) Move forgot password logic to http-server
- This should simplify the overall interaction and removes logic from the complicated ServerQuarantineConversation.
- The forgot password logic was in the lobby code for a few reasons:
  - was not clear how or if emails could be sent from http server
  - was questionable how and if password hashing could be re-used.
- Email was previously not part of the lobby socket communication as adding that additional piece of information was not compatible with the previous flow. It was not readily apparent if it could have been cleanly added, so no email was at the time a limitation of re-using LobbyServerQuarantineConversation. With the logic moved to HTTP Server, it's very straight forward to add email to the request JSON being sent to the server.

(3) Minor: In non-prod we continue to *not* send a temp password email, but we will now log it.  Previously the email notificationit was a complete no-op and there was no way to know the temp password (in non-prod).

(4) Stop looking at XFORWARDED_HEADER when looking for IP address in http requests. It can be forged, if someone is using a Proxy we'll read that IP address and its okay.