Add shift+d hotkey to close purchase panel
This update adds shift+d hotkey to close purchase panel (clicks done button).

While purchase panel has multiple ways to close, the 'done' button is not bound to a hotkey.
On the action paneles, the 'done' buttons are bound to 'shift+d' on each panel, but not
so for the purchase panel. Ending tech research for example with 'shift-d' to then see
purchase panel and a 'done' button on purchase that cannot be activated with 'shift+d'
seems inconsistent. After this update the purchase panel can be closed with shift+d