Update travis to delete existing tag if it exists
Potentially a build fix, it looks like introducing 'set -eu' is failing the build. Previously the build would blindly try to delete tags and we ignored any failures there. With 'set -eu', this potentially causes a failure.

The update here adds a 'git ls-remote --tags' to look for tags and then conditionally executes the delete tag if the target tag is found.