Fix pre-release to release bad enum value problem for flag settings
Newest release sets an enum value that is not present in previous release. When the previous release reads that value, it errors out with 'enum value not found'. To fix this, update latest to null out the enum flag value (even though it is not necessary). This way when flags are turned off and a null value is present, the latest release will read a null value and would not error out.

Reported in: https://forums.triplea-game.org/topic/1531/menu-absent-new-bug-in-old-version/5