Parallel travis VMs
Break up build tasks across a couple more Travis VMs by using build matrix. Should help parallelize build tasks and allow for a faster overall runtime of the build.

Release task is broken out and pre-emptively downloads images while we expect other tasks to still be running. This way if we get to the deploy step (merge builds), then the images will already be downloaded and we can proceed with artifact creation.

Five parallel jobs is chosen specifically as that seems to be the max number of VMs that Travis will spin up for a project fork. Beyond that and we have jobs waiting for previous jobs to terminate to free up VMs. Hence, five seems to be the maximum parallelization available from travis.

Tasks load balanced in an attempt that each of the 5 VMs will take the same amount of time.