Update version compatibility algorithm
Update version number to be interpretted as "<compat>.<release>.<build>"

Two versions are compatible if the leading digit is equal.
The second number is to indicate releases, which may be compatible with previous releases.
The point number, or build number, is assigned by the build system and is ever incrementing.

For map compatibility, each version number except the point number is looked at, the compatibility
number of maps is currently both 'compat' and 'release' numbers.