Travis build stages / Parallel Builds
(1) Configure travis to use build stages. This allows for the 'verify' stage to be done in parallel. Note, each stage runs on its own VM. After the 'verify' stage is done, if we are building master, we'll launch a deploy stage. After all the 'verify' VMs complete, the 'deploy' stage will kick off if we are building master. Since each stage uses its own VM, there is no interaction between the deploy step and previous steps.

(2) Add 'on tag is missing' to skip tag builds. This should avoid dupliate builds after merges to master as the tag that is pushed to trigger deployments triggers its own (redundant) build.

Note: the 'env' variable provided is only for descriptive purposes and is displayed in the UI to help label each build job.

Going to stages we see a number of benefits:
- no need for variables, the deploy job can be explicitly run for just JDK11
- avoiding build matrix is a bit simpler, we can instead explicitly indicate which JDK to use for which build stage
- avoids redundant setup and unnecessary resources. For example we can setup a DB only on the stage that needs it, install python dependencies to only the VM that needs python (for map yaml splitting), and can run 'setup_gpg' for only the deployment stage which is the only step that needs it.
- parallelism, jobs typically finish in under 9 minutes running across 4 VMs