Squash changesets together
Because we are doing a clean DB install, all changesets are being squashed together. If that were not the case, then all of the V1.10 changesets would be squashed together only and leave the already applied V1.00.00__lobby.sql changeset unsquashed.

There are no changes other than concatenating files together.