Remove Checkstyle summary javadoc rule (#5250)
The rule requires a first sentence to end in a period.

This has resulted in 'hack' situations where the first sentence ends naturally in a colon,
in which case an extraneous period is introduced ".:"

More fundamentally it's useful/important to be required to add meaningful javadocs, otherwise
we likely won't add them. With that, this syntax requirement is not that helpful and
gets into a nitpicking territory that we do not enforce at review, ie: it's not
important for a single sentence javadoc to end in a period or not.