Use 'showChatTime' client setting over property
- Makes the lobby 'showChatTime' selection persistent, uses client setting now.
- Hide 'showChatTime' menu option if there is no chat rather than disabling it.

'showChatTime' existed in both lobby menu and the game menu. The game menu was updated to use the client setting, this update changes the lobby menu to do the same. Of interest as well the headless chat used this property even though headless chat would never be able to set it, it is up to the client to prepend chat time.