Refactor: Convert 'String' to value object 'PlayerName'
First iteration to mass convert unbounded strings to domain object 'PlayerName'.
Refactor begins with updating the API of of 'Node' to have a 'getPlayerName'
method with objective to eventually to convert the internal data type of
'Node.name' to be 'PlayerName' and be strongly typed whenever accessed.

Another goal behind this conversion is to avoid passing 'Node' data where
typically only one property of the Node, name, is used.