Fix: Headless bot select player option
To show a 'select' player button, we need to pass a null value to the
checker method. Updating to value objects, we instead had a null
value wrapped in a non-null method, which the checker method then
did not deal with properly.

This update fixes the situation by restoring previous behavior where
null player names are passed as null. It is tempting to allow the
value object 'PlayerName' to simply hold a null (so we could check
if the contained value is null rather than checking teh value object),
but for now leaning towards future-behavior where we would
enforce non-null values to be held by PlayerName.