Refactor: Split Chat from Communication Mechanism
Main thrust of this update is to split the communication mechanism
from chat. This is not a complete update but moves a long way, future
updates should be able to drop in new 'ChatTransmitter' implementations
to allow for chat to be backed by a different transport technology. In
other words, this update attempts to 'hide' messengers interaction from
Chat

Notable updates:
- IP address of chatters is (more) hidden

This update also tries to hide/remove 'node' from Chat usage. Of
significant, the front-end should have fewer if any node references.
This was a security issue as node carries IP address with it, before
this update the IP addresses of all chat participants is broadcasted
to all other chat clients.

- StatusController squashed into ChatController

Mainly a simplification change, reduces number of remotes.

- ChatParticpant data object introduced, represents a player name
  plus tag (moderator or not moderator).

Change summary:
- Remove INode from Chat, use ChatParticipant instead
- Move 'messengers' interaction out of Chat to be behind 'ChatTransmitter'
- Squash 'StatusController' into 'ChatController'
- Update status data representation, instead of parallel maps
  of "player -> tag" and "player -> status", create a data
  object "ChatParticipant" that is a combination of name and
  tag, and change data structure to a single map "chatparticipant -> status"
- Update headless chat 'setChat' method to be an unsupported operation.
  Headless chat is instantiated just once and it appears the cancel
  path is never invoked to unset the chat. Of note, the 'history' copy
  that was done in the 'setChat' method can be removed as well since
  we only 'set' with a new chat object, once, the history is always
  empty.
- Use empty status to represent no status instead of null.
- Remove chat version numbering, was not the best guard for ordering
  and is strictly not necessary
- Remove chat latch and synchronization, latch is unnecessary and
  synchronization is not needed or would have wrapped what are thread-safe
  or atomic operations.