Refactor: Remove Observable from SetupModel
(1) Remove observer methods from SetupModel, this cascades
to HeadlessServerSetup and turns out we add no observers
to HeadlessServerSetup. Looking at the headed version, the
observe methods are used to update UI, it makes sense then
that headless is a no-op.

(2) Convert SetupPanel Observer interface to a listener with
a Java8 API.