Refactor: HttpController base class and ErrorReportController consistency (#5302)
* Refactor: HttpController base class and ErrorReportController consistency

1) Add a base class for HttpControllers called HttpController. Base classes
tend to be anti-patterns, except it seems to work well when the base class is
little more than the common annotations. It's easy to forget to add the @Path("")
annotation, by offering a common class to extend, with luck that detail will be
easier to include and also the class will provide a location for documentation
on how to build an endpoint.

2) ErrorUpload* was inconsistently named from its controller and was renamed ErrorReport*
It was also missing a test, the rename and addition of test case brings it in-line to
the standards of the other controller+client classes.