Refactor: Move connectivity check to http
Migrate connectivity test to http server from java socket server.
The connectivity check is the verification that kicks in after hosting
a game where the server tries to establish a 'reverse' connection
to the host to make sure the host is accessible from the internet.

This update moves that check to HTTP server.