Major Backend Update: Use Http Lobby for GameListing
Overview:

This update removes the Java NIO lobby backend code for managing games and instead
uses the http server backend. Most of the updates are in lobby watcher where
game hosts interact with the lobby to send updates. The other main area of update
is in the lobby game table model which now polls for game updates and calls
existing game update listener methods when there is a game update.

Of note, a new endpoint and client were added: GameHostingController + Client. These
provide an endpoint where bots and game hosts can request an API key without prior
authentication. We need this as bots will not go through the standard login process.
We previously stored API key with username, the new endpoint and code do not accept
or require a username.

Update list:
- Potentially temporary remove of check for already logged in player and name de-duping.
  The check prevented connecting to a hosted game multiple times and needs some re-work.
- Remove backend Java NIO components for lobby game listing: ILobbyGameController and
  ILobbyGameBroadcaster
- Change GameId from UUID to String. This is so we have looser coupling on data types
  and do not require the game id to actually be of type UUID.
- Remove 'assertFromServer' in LobbyTableModel, messages can only be from server
- Remove hostname property from GameDescription, instead get the same value from the
  hostNode property
- Remove mutex in lobby watcher thread, the synchronization does not help us as network
  messages could be re-ordered
- Swap ArrayList in SetupPanel for copyOnWrite to avoid concurrent modification error.
- Delete restarting of lobby watcher thread, lobby watching background thread is done
  from the lobby watcher itself and handles communication errors, it should not need
  periodic restarts.
- Change bot scripts to use https port only, bots no longer use the Java NIO lobby port.
- Add Timers utility class to provide a convenient type-safe builder API for creating
  recurring tasks.

Notable TODOs:
- Moderator boot game option is broken
- Host filtering to disallow banned users from hosting
- Combine connectivity check with 'post-game' controller and allow for custom hostname