Send bot messages line by line, restore message cap to hosts
Main goal is to remove the 'serverPlayer' property from Chat.java
which is only used in ChatMessagePanel to determine if messages
are coming from the server player in order to not truncate them.

This is done as bots send stack traces to chat and to avoid
truncating those chat messages. It can be the case though
that a server player is a human, not bot, in which case they
are the one human player without a chat limit.

This update restores the message cap to all players and to fix
the bot stack trace situation, stack traces are sent line-by-line.
This is a possible solution now because flood control has been
removed from bots. Flood control is done client-side, which did
not make sense for bots and previously would have prevented
this situation. Client-side flood control does not make sense
for bots and was removed, which enables this solution.