Add domain-data subproject
New sub-project for domain-entities that are heavily shared between multiple, potentially low-level sub-projects.
For example, at the DB layer, it makes sense to have API's to accept 'PlayerName', it has become a bit untenable
to keep the DB layer to pure primitive java entities. It is also getting very odd that common domain objects are
converging to 'http-clients' (where they do not belong).

It is a risk we'll add too many or inappropriate domain objects to the new sub-project, we should be judicious
about having only true domain and heavily shared objects land in 'domain-data'. Specifically, with multiple
sub-domains, a term can very well be defined in multiple sub-domains and have different meanings. If we place
such objects in 'domain-data', the multiple means will conflict, this is something to avoid.