Add ClientMessageEnvelope & Factory
Add objects that will be sent from client to server over websocket.
The message envelope is designed to be encoded into a JSON string
and contains three fields. The envelope contains a message type and
payload. Based on message type the server will know what kind of object
exists in the payload. Last, the message envelope contains the clients
API-key to be used by the server for client identification and filtering.