Wire LobbyChatTransmitter to LobbyChatClient
Adds client side infrastructure for lobby websocket chat.
LobbyChatTransmitter is a new class that can be swapped out for a
'MessengersChatTransmitters' as they both implement the same chat API.

The ChatTransmitter is wired to a set of listener APIs and send
APIs that are backed by an http component that eventually sends
and receives requests from the GenericWebsocket connection.

As part of this wiring is an event handler that handles server
messages and routes them back to the appropriate listener.